Bdcalling chatbot Fast API/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
typing-extensions
